federation-api: Add support for partial state in send_join
According to MSC3706
