+++ crates/ruma-federation-api/Cargo.toml
@@ -23,6 +23,7 @@ unstable-exhaustive-types = []
 unstable-msc2448 = []
 unstable-msc3030 = []
 unstable-msc3618 = []
+unstable-msc3706 = []
 unstable-msc3723 = []
 unstable-unspecified = []
