+++ crates/ruma/Cargo.toml
@@ -153,6 +153,7 @@ unstable-msc3553 = ["ruma-common/unstable-msc3553"]
 unstable-msc3554 = ["ruma-common/unstable-msc3554"]
 unstable-msc3575 = ["ruma-client-api?/unstable-msc3575"]
 unstable-msc3618 = ["ruma-federation-api?/unstable-msc3618"]
+unstable-msc3706 = ["ruma-federation-api?/unstable-msc3706"]
 unstable-msc3723 = ["ruma-federation-api?/unstable-msc3723"]
 unstable-msc3783 = ["ruma-common/unstable-msc3783"]
 unstable-msc3931 = ["ruma-common/unstable-msc3931"]
@@ -193,6 +194,7 @@ __ci = [
     "unstable-msc3554",
     "unstable-msc3575",
     "unstable-msc3618",
+    "unstable-msc3706",
     "unstable-msc3723",
     "unstable-msc3783",
     "unstable-msc3932",
